Make deployment script run portably
The deployment script was hard-coded to the original workspace
environment and needed to be made portable for other locations.

The solution replaced hard-coded path prefix and instead determined
it at runtime. That value was used to set the local deployment
directory, which is used as the mount volume for the Docker
deployment instance.