Bug #3 Docker running Payara reports SEVERE messages
After looking at the log messages it appeared the other files and
directories in /target where being read by Payara's autodeployment
and throwing errors for not being a WAR file. The best solution
would be to isolate the target Hello.war file to it's own deployment
directory without any extranious files.

The deployment script was modified to manage the cleaning, creation,
and copying of the WAR file into a dedicated deployment directory.
Comments were also added to the script, explaining what it does.