Feature #5 Expose SSL to application
The Payara server runs the application over HTTP and HTTPS (8080/8181).
The deploy script only mapped the HTTP port to the localhost. To view
the application under SSL, the deploy script was modified to expose
port 8181 to the localhost. The script's documentation was updated to
reflect this change.

Additionally, the README.md was updated to better reflect the current
usage of the application.